pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptgroundboxbench"
version = "0.1.0"
description = "Prompt driven bounding box benchmark: Grounding DINO vs YOLO11."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "PromptGroundBoxBench Authors" }]
dependencies = [
  "typer>=0.12",
  "PyYAML>=6.0",
  "numpy>=1.24",
  "Pillow>=10.0",
  "rich>=13.7",
  "tqdm>=4.66",
]

[project.optional-dependencies]
# Keep torch optional so CI can run quickly. Install torch separately for your GPU.
torch = [
  "torch>=2.1",
  "torchvision>=0.16",
]
inference = [
  "transformers>=4.43",
  "accelerate>=0.33",
  "ultralytics>=8.3",
]
demo = [
  "gradio>=4.0",
]
eval = [
  "pycocotools>=2.0.7",
]
all = [
  "torch>=2.1",
  "torchvision>=0.16",
  "transformers>=4.43",
  "accelerate>=0.33",
  "ultralytics>=8.3",
  "gradio>=4.0",
  "pycocotools>=2.0.7",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
  "types-PyYAML",
]

[project.scripts]
pgbb = "promptgroundboxbench.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
ignore = [
  "E501", # line length handled pragmatically
]

[tool.ruff.lint.isort]
known-first-party = ["promptgroundboxbench"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
